move the rendering of the table headers to the CohortView component and keep content to fill the table in the CohortFilter table. Having issues getting the state of currCohort to be set prior to render which prevents filter functionality
