+++ components/AdminPage/index.js
@@ -8,8 +8,10 @@ import RevealChat from './RevealChat'
 const AdminContainer = ({ allCohorts }) => {
   const [cohorts] = useState(allCohorts);
   const [currCohort, setCurrCohort] = useState([]);
+   const [studentData, setStudentData]= useState([])
   const [menuClicked, setMenuClicked] = useState(false)
   const [chatCohort,setChatCohort] = useState('')
+
   useEffect(() => {
     (async () => {
       if (cohorts.length > 0) {
@@ -24,6 +26,8 @@ const AdminContainer = ({ allCohorts }) => {
             students,
           },
         ]);
+        //setStudentData([...students])
+        console.log("adminpage currCohort:", currCohort)
       }
     })();
   }, []);
@@ -54,13 +58,13 @@ const AdminContainer = ({ allCohorts }) => {
       <div className={s.container}>
         <div className={s.tools_container}>
          <div >
-            <CohortMenu setChatCohort={setChatCohort} toggleMoveChat={toggleMoveChat} cohorts={cohorts} currCohort={currCohort} setCurrCohort={setCurrCohort}  />
+            <CohortMenu setStudentData={setStudentData} setChatCohort={setChatCohort} toggleMoveChat={toggleMoveChat} cohorts={cohorts} currCohort={currCohort} setCurrCohort={setCurrCohort}  />
          </div>
          <motion.div initial="enter" animate={menuClicked ? "exit" : "enter" } variants={moveMenuAnimate}> 
           <RevealChat chatCohort={chatCohort}/>
          </motion.div>
         </div>
-        <CohortView setCurrCohort={setCurrCohort} currCohort={currCohort} setChatCohort={setChatCohort} />
+        <CohortView studentData={studentData} setCurrCohort={setCurrCohort} currCohort={currCohort} setChatCohort={setChatCohort} />
     </div>
   </div>
   )
